style: 대시보드 차트 수정 (#9)
