+++ src/shared/components/admin/access-status/access-status-chart-section.tsx
@@ -1,5 +1,5 @@
 import { useAccessStatusChart } from '@hooks/use-access-status-chart';
-import { AccessStatusChartSectionView } from '@components/admin/access-status/access-status-chart-view';
+import AccessStatusChartSectionView from '@components/admin/access-status/access-status-chart-view';
 
 const AccessStatusChartSection = () => {
   const chartState = useAccessStatusChart();
