+++ src/shared/routes/admin-guard.tsx
@@ -1,11 +1,16 @@
 import { useAtomValue } from 'jotai';
 import { Navigate, Outlet, useLocation } from 'react-router-dom';
-import { isAdminAtom } from '@state/session';
+import { getStoredUserRole, isAdminAtom } from '@state/session';
 
 const AdminGuard = () => {
   const isAdmin = useAtomValue(isAdminAtom);
   const loc = useLocation();
-  if (!isAdmin) return <Navigate to="/forbidden" replace state={{ from: loc }} />;
+  const storedRole = getStoredUserRole();
+
+  if (!isAdmin && storedRole !== 'ADMIN') {
+    return <Navigate to="/forbidden" replace state={{ from: loc }} />;
+  }
+
   return <Outlet />;
 };
