+++ src/shared/routes/auth-guard.tsx
@@ -1,11 +1,16 @@
 import { useAtomValue } from 'jotai';
 import { Navigate, Outlet, useLocation } from 'react-router-dom';
-import { isLoggedInAtom } from '@state/session';
+import { getStoredAccessToken, isLoggedInAtom } from '@state/session';
 
 const AuthGuard = () => {
   const isLoggedIn = useAtomValue(isLoggedInAtom);
   const loc = useLocation();
-  if (!isLoggedIn) return <Navigate to="/login" replace state={{ from: loc }} />;
+  const hasStoredToken = Boolean(getStoredAccessToken());
+
+  if (!isLoggedIn && !hasStoredToken) {
+    return <Navigate to="/login" replace state={{ from: loc }} />;
+  }
+
   return <Outlet />;
 };
